SparseDC simulation

Simulate counts from cluster in two conditions using the SparseDC method.

sparseDCSimulate(params = newSparseDCParams(), verbose = TRUE, ...)

Arguments

params
SparseDCParams object containing simulation parameters.

verbose
logical. Whether to print progress messages

...
any additional parameter settings to override what is provided in params.

Value

SingleCellExperiment containing simulated counts

Details

This function is just a wrapper around sim_data that takes a SparseDCParams, runs the simulation then converts the output from log-expression to counts and returns a SingleCellExperiment object. The original simulated log-expression values are returned in the LogExprs assay. See sim_data and the SparseDC paper for more details about how the simulation works.

References

Campbell K, Yau C. Uncovering genomic trajectories with heterogeneous genetic and environmental backgrounds across single-cells and populations. bioRxiv (2017).

Barron M, Zhang S, Li J. A sparse differential clustering algorithm for tracing cell type changes via single-cell RNA-sequencing data. Nucleic Acids Research (2017).

Paper: 10.1093/nar/gkx1113

Examples

if (requireNamespace("SparseDC", quietly = TRUE)) {
    sim <- sparseDCSimulate()
}

#> Simulating counts...

#> Creating final dataset...