splatter: Simple Simulation of Single-cell RNA Sequencing Data

Splatter is a package for the simulation of single-cell RNA sequencing count data. It provides a simple interface for creating complex simulations that are reproducible and well-documented. Parameters can be estimated from real data and functions are provided for comparing real and simulated datasets.

See also

Author

Maintainer: Luke Zappia luke@lazappi.id.au (ORCID) (lazappi)

Authors:

Belinda Phipson belinda.phipson@petermac.org (ORCID) (bphipson)

Alicia Oshlack alicia.oshlack@petermac.org (ORCID)

Other contributors:

Christina Azodi cazodi@svi.edu.au (ORCID) (azodichr) [contributor]