scDD simulation

Simulate counts using the scDD method.

scDDSimulate(params = newSCDDParams(), plots = FALSE,
  plot.file = NULL, verbose = TRUE, BPPARAM = SerialParam(), ...)

Arguments

params
SCDDParams object containing simulation parameters.

plots
logical. whether to generate scDD fold change and validation plots.

plot.file
File path to save plots as PDF.

verbose
logical. Whether to print progress messages

BPPARAM
A BiocParallelParam instance giving the parallel back-end to be used. Default is SerialParam which uses a single core.

...
any additional parameter settings to override what is provided in params.

Value

SingleCellExperiment containing simulated counts

Details

This function is just a wrapper around simulateSet that takes a SCDDParams, runs the simulation then converts the output to a SingleCellExperiment object. See simulateSet for more details about how the simulation works.

References

Korthauer KD, Chu L-F, Newton MA, Li Y, Thomson J, Stewart R, et al. A statistical approach for identifying differential distributions in single-cell RNA-seq experiments. Genome Biology (2016).

Paper: 10.1186/s13059-016-1077-y

Code: https://github.com/kdkorthauer/scDD

Examples

# NOT RUN {
sim <- scDDSimulate()
# }