Estimate Splat simulation parameters

Estimate simulation parameters for the Splat simulation from a real dataset. See the individual estimation functions for more details on how this is done.

splatEstimate(counts, params = newSplatParams())

# S3 method for SingleCellExperiment
splatEstimate(counts,
  params = newSplatParams())

# S3 method for matrix
splatEstimate(counts, params = newSplatParams())

Arguments

counts
either a counts matrix or a SingleCellExperiment object containing count data to estimate parameters from.

params
SplatParams object to store estimated values in.

Value

SplatParams object containing the estimated parameters.

See also

splatEstMean, splatEstLib, splatEstOutlier, splatEstBCV, splatEstDropout

Examples

# Load example data
library(scater)
data("sc_example_counts")

params <- splatEstimate(sc_example_counts)
params

#> A Params object of class SplatParams 
#> Parameters can be (estimable) or [not estimable], 'Default' or 'NOT DEFAULT'. 
#> 
#> Global: 
#> (GENES)  (CELLS)   [Seed]  
#>    2000       40   578513  
#> 
#> 27 additional parameters 
#> 
#> Batches: 
#>     [Batches]  [BATCH CELLS]     [Location]        [Scale]  
#>             1             40            0.1            0.1  
#> 
#> Mean: 
#>             (RATE)             (SHAPE)  
#> 0.0032639148887191   0.419304527025529  
#> 
#> Library size: 
#>       (LOCATION)           (SCALE)  
#> 12.7053037408488  0.58910919018286  
#> 
#> Exprs outliers: 
#>       (PROBABILITY)           (LOCATION)              (SCALE)  
#> 0.00623376623376623     4.22280669039478    0.422108697251885  
#> 
#> Groups: 
#>      [Groups]  [Group Probs]  
#>             1              1  
#> 
#> Diff expr: 
#> [Probability]    [Down Prob]     [Location]        [Scale]  
#>           0.1            0.5            0.1            0.4  
#> 
#> BCV: 
#>    (COMMON DISP)             (DOF)  
#> 3.21969047393022  14.7262017878812  
#> 
#> Dropout: 
#>          [Present]          (MIDPOINT)             (SHAPE)  
#>              FALSE    5.00420976858168  -0.652889139821156  
#> 
#> Paths: 
#>         [From]        [Length]          [Skew]    [Non-linear]  [Sigma Factor]  
#>              0             100             0.5             0.1             0.8  
#>