Estimate Splat simulation parameters

Estimate simulation parameters for the Splat simulation from a real dataset. See the individual estimation functions for more details on how this is done.

splatEstimate(counts, params = newSplatParams())

# S3 method for SingleCellExperiment
splatEstimate(counts, params = newSplatParams())

# S3 method for matrix
splatEstimate(counts, params = newSplatParams())

Arguments

counts
either a counts matrix or a SingleCellExperiment object containing count data to estimate parameters from.

params
SplatParams object to store estimated values in.

Value

SplatParams object containing the estimated parameters.

See also

Examples

# Load example data
library(scater)
set.seed(1)
sce <- mockSCE()

params <- splatEstimate(sce)

#> NOTE: Library sizes have been found to be normally distributed instead of log-normal. You may want to check this is correct.

params

#> A Params object of class SplatParams 
#> Parameters can be (estimable) or [not estimable], 'Default' or  'NOT DEFAULT' 
#> Secondary parameters are usually set during simulation
#> 
#> Global: 
#> (GENES)  (CELLS)   [Seed] 
#>    2000      200   374097 
#> 
#> 28 additional parameters 
#> 
#> Batches: 
#>     [BATCHES]  [BATCH CELLS]     [Location]        [Scale] 
#>             1            200            0.1            0.1 
#> 
#> Mean: 
#>            (RATE)            (SHAPE) 
#> 0.002962686167343  0.496997730070513 
#> 
#> Library size: 
#>       (LOCATION)           (SCALE)            (NORM) 
#>       357331.235  11607.2332293176              TRUE 
#> 
#> Exprs outliers: 
#> (PROBABILITY)     (Location)        (Scale) 
#>             0              4            0.5 
#> 
#> Groups: 
#>      [Groups]  [Group Probs] 
#>             1              1 
#> 
#> Diff expr: 
#> [Probability]    [Down Prob]     [Location]        [Scale] 
#>           0.1            0.5            0.1            0.4 
#> 
#> BCV: 
#>     (COMMON DISP)              (DOF) 
#> 0.752043426792845   11211.8933424157 
#> 
#> Dropout: 
#>            [Type]         (MIDPOINT)            (SHAPE) 
#>              none   2.71153535179343  -1.37209356733765 
#> 
#> Paths: 
#>         [From]         [Steps]          [Skew]    [Non-linear]  [Sigma Factor] 
#>              0             100             0.5             0.1             0.8 
#>